refactor(adapters-cursor): isolate desktop executor type contract (split S04 L0/5)
